+++ inc/HM11/hm11.c
@@ -1,6 +1,25 @@
 #include "hm11.h"
+#include "../MPU6050/mpu6050.h"
 
-static void Hm11_Init(void){
+void Hm11_Packet(void){
+    _tx_buffer_[0] = Mpu6050_.Accel_X_RAW >> 8;
+    _tx_buffer_[1] = Mpu6050_.Accel_X_RAW;
+    _tx_buffer_[2] = Mpu6050_.Accel_Y_RAW >> 8;
+    _tx_buffer_[3] = Mpu6050_.Accel_Y_RAW;
+    _tx_buffer_[4] = Mpu6050_.Accel_Z_RAW >> 8;
+    _tx_buffer_[5] = Mpu6050_.Accel_Z_RAW;
+    _tx_buffer_[6] = Mpu6050_.Gyro_X_RAW >> 8;
+    _tx_buffer_[7] = Mpu6050_.Gyro_X_RAW;
+    _tx_buffer_[8] = Mpu6050_.Gyro_Y_RAW >> 8;
+    _tx_buffer_[9] = Mpu6050_.Gyro_Y_RAW;
+    _tx_buffer_[10] = Mpu6050_.Gyro_Z_RAW >> 8;
+    _tx_buffer_[11] = Mpu6050_.Gyro_Z_RAW;
+//    _tx_buffer_[12] = Mpu6050_.Temperature >> 8;
+//    _tx_buffer_[13] = Mpu6050_.Temperature;
+//    _tx_buffer_[14] = Mpu6050_Config_.gyro_sensitivity << 4 | Mpu6050_Config_.accel_sensitivity;
+}
+
+void Hm11_Init(void){
     /* ---------- UART Initialize ---------- */
     /* Initialize CoreUARTapb with its base address, baud value, and line configuration */
     UART_init( &g_uart_0, COREUARTAPB0_BASE_ADDR,