+++ inc/MPU6050/corei2c_driver.h
@@ -1,7 +1,7 @@
-#ifndef COREI2C_DRIVER_H_
-#define COREI2C_DRIVER_H_
-
-#include "../../firmware/drivers/CoreI2C/core_i2c.h"
+//#ifndef COREI2C_DRIVER_H_
+//#define COREI2C_DRIVER_H_
+//
+//#include "../../firmware/drivers/CoreI2C/core_i2c.h"
 
 // static void Mpu6050_Init(void){
 //     // MPU6050 fast mode is 400kHz std mode is 100 kHz
@@ -9,4 +9,4 @@
 
 // }
 
-#endif
+//#endif